move user database to settings
