+++ mex/drop/models/user.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class User(BaseModel):
+    """User model."""
+
+    username: str
+    x_system: str
