+++ mex/drop/security.py
@@ -2,35 +2,13 @@ from typing import Annotated
 
 from fastapi import Depends, HTTPException
 from fastapi.security import APIKeyHeader
-from pydantic import BaseModel
 from starlette import status
 
+from mex.drop.models.user import User
+from mex.drop.settings import DropSettings
+from mex.drop.types.user_database import UserDatabase
+
 X_API_KEY = APIKeyHeader(name="X-API-Key")
-
-
-class User(BaseModel):
-    """User model."""
-
-    username: str
-    x_system: str
-
-
-UserDatabase = dict[str, User]
-
-
-FAKE_USERS_DB: UserDatabase = {
-    k: User(**v)
-    for k, v in {
-        "johndoe": {
-            "username": "johndoe",
-            "x_system": "test_system",
-        },
-        "alice": {
-            "username": "alice",
-            "x_system": "foo_system",
-        },
-    }.items()
-}
 
 
 def get_user(db: UserDatabase, username: str) -> User | None:
@@ -38,15 +16,12 @@ def get_user(db: UserDatabase, username: str) -> User | None:
 
     Args:
         db: database dictionary
-        username: user name
+        username: username
 
     Returns:
         User if username in db, else None
     """
-    if username in db:
-        return db[username]
-    else:
-        return None
+    return db.get(username)
 
 
 def get_current_user(api_key: Annotated[str, Depends(X_API_KEY)]) -> User:
@@ -57,10 +32,16 @@ def get_current_user(api_key: Annotated[str, Depends(X_API_KEY)]) -> User:
     Args:
         api_key: the API key for user lookup
 
+    Settings:
+        drop_user_database: checked for presence of api_key
+
     Returns:
         User
     """
-    user = get_user(FAKE_USERS_DB, api_key)
+    settings = DropSettings.get()
+    user_db = settings.drop_user_database
+
+    user = get_user(user_db, api_key)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
