+++ mex/drop/types/user_database.py
@@ -1,0 +1,3 @@
+from mex.drop.models.user import User
+
+UserDatabase = dict[str, User]
