+++ tests/conftest.py
@@ -4,9 +4,22 @@ import pytest
 from fastapi.testclient import TestClient
 
 from mex.drop.main import app
+from mex.drop.models.user import User
 from mex.drop.settings import DropSettings
 
 pytest_plugins = ("mex.common.testing.plugin",)
+
+
+TEST_USER_DATABASE = {
+    "johndoe": {
+        "username": "johndoe",
+        "x_system": "test_system",
+    },
+    "alice": {
+        "username": "alice",
+        "x_system": "foo_system",
+    },
+}
 
 
 @pytest.fixture(autouse=True)
@@ -14,6 +27,7 @@ def settings(tmp_path: Path) -> DropSettings:
     """Load the settings for this pytest session."""
     settings = DropSettings.get()
     settings.drop_root_path = str(tmp_path)
+    settings.drop_user_database = {k: User(**v) for k, v in TEST_USER_DATABASE.items()}
     return settings
